requirements.txt
azure-cognitiveservices-speech
pydub
python-dotenv
grpc
grpcio 
grpcio-tools
os
groq